Relocated App to components folder
