+++ src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './styles/index.css';
-import App from './App';
+import App from './components/App/App';
 
 ReactDOM.render(
   <React.StrictMode>
